initial ubi8/s2i-base support
This commit uses ubi8/s2i-base as the base image. It currently uses rpms
(nodejs/npm) that are in this repository. This is due to some issues we
are having building them. When those issues are resolved the rpms will
be removed and installed from a remote location published by node-rpm.